pyproject.toml
[tool.poetry]
authors = ["m09 <142691+m09@users.noreply.github.com>"]
description = ""
name = "tp-compression"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
dvc = "^3.43.1"
dvclive = "^3.41.1"
numpy = "^1.26.3"
opencv-python = "^4.9.0.80"
python = ">= 3.10, < 3.11"
tensorflow = "<2.15"
scikit-learn = "^1.4.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pylint = "^3.0.3"
flake8 = "^7.0.0"
black = "^24.1.1"
types-pyyaml = "^6.0.12.12"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
